+++ Source/NexusForever.Network.World/Message/Model/Reputation/ServerReputationUpdate.cs
@@ -1,7 +1,7 @@
 ﻿using NexusForever.Game.Static.Reputation;
 using NexusForever.Network.Message;
 
-namespace NexusForever.Network.World.Message.Model
+namespace NexusForever.Network.World.Message.Model.Reputation
 {
     [Message(GameMessageOpcode.ServerReputationUpdate)]
     public class ServerReputationUpdate : IWritable
