+++ Source/NexusForever.Network/Message/GameMessageOpcode.cs
@@ -559,6 +559,7 @@ namespace NexusForever.Network.Message
         ServerVehiclePassengerAdd       = 0x086F,
         ServerEntityAggroSwitch         = 0x0876,
         ServerEntityCCStateSet          = 0x087F,
+        ServerReputationOverrideRemove  = 0x0885,
         ServerEntityDeathState          = 0x088C,
         ServerDuelCountdown             = 0x0893,
         ServerDuelResult                = 0x0894,
@@ -576,6 +577,7 @@ namespace NexusForever.Network.Message
         ServerUnitPvpStateChange        = 0x08BD,
         ServerVehiclePassengerRemove    = 0x08C7,
         ServerUnitSetChair              = 0x08CF,
+        ServerReputationOverrideAdd     = 0x08D7,
         ServerEntityVisualUpdate        = 0x0905,
         ServerEntityTargetUnit          = 0x0908,
         ServerEntityThreatListUpdate    = 0x0909,
